Admin: move code for AdminNotices component to its own file.
